% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRP_Portfolio.R
\name{HRP_Portfolio}
\alias{HRP_Portfolio}
\title{Hierarchical Risk Parity portfolio}
\usage{
HRP_Portfolio(covar, clustering.method = "single")
}
\arguments{
\item{covar}{Covariance matrix of returns. The covariance matrix will be transformed into correlation matrix and then into distance matrix.}

\item{clustering.method}{Linkage method used in the hierarchical clustering. Allowed options are `single`, `complete`, `average` or `ward`. Default option is `single`.}
}
\value{
portfolio weights
}
\description{
Performs the Hierarchical Risk Parity portfolio proposed strategy by De Prado (2016). Several linkage methods for the hierarchical clustering can be used, by default the `single` linkage is used.
}
\seealso{
`HCAA_porfolio` and `HERC_Portfolio`
}
