\name{mldp_returns}
\alias{mldp_returns}
\title{
Returns of 10 simulated assets.
}
\description{
This dataset contains the simulated returns used in the numerical example of Marcos Lopez de Prado's paper, hence the name \code{mldp_returns}. The Python code used to reproduce this simulated data is kindly provided by the author in the supplementary material of his paper.
}
\references{
De Prado, Marcos Lopez. "Building diversified portfolios that outperform out of sample." The Journal of Portfolio Management 42.4 (2016): 59-69.
}
\keyword{dataset}
